requirements.txt
swbs
pycryptodomex
pyserial
numpy
opencv-contrib-python
imutils
sense-hat
gpiozero
